feat: adjust i2s usable register values
