+++ src/registers/mod.rs
@@ -1,5 +1,6 @@
 //! Register modules and address definitions.
 
+pub mod all_int_src;
 pub mod counter_bdr;
 pub mod ctrl10_c;
 pub mod ctrl1_xl;
@@ -11,31 +12,35 @@ pub mod ctrl6_c;
 pub mod ctrl7_g;
 pub mod ctrl8_xl;
 pub mod ctrl9_xl;
-pub mod ctrl_ois;
+pub mod drd_src;
 pub mod emb_func_status;
 pub mod fifo_ctrl;
 pub mod fifo_data_out;
+pub mod fifo_data_out_tag;
 pub mod fifo_status;
 pub mod free_fall;
+pub mod fsm_status;
 pub mod func_cfg_access;
 pub mod int_ctrl;
 pub mod int_dur2;
-pub mod int_ois;
-pub mod int_src;
 pub mod internal_freq_fine;
 pub mod md_cfg;
+pub mod mlc_status;
 pub mod ofs_usr;
-pub mod out_data;
+pub mod out_temp;
 pub mod pin_ctrl;
+pub mod status_master;
 pub mod status_reg;
 pub mod tap_cfg;
+pub mod tap_src;
 pub mod tap_ths_6d;
 pub mod timestamp;
 pub mod wake_up_dur;
+pub mod wake_up_src;
 pub mod wake_up_ths;
 
+pub use all_int_src::*;
 pub use counter_bdr::*;
-pub use ctrl_ois::*;
 pub use ctrl1_xl::*;
 pub use ctrl2_g::*;
 pub use ctrl3_c::*;
@@ -46,26 +51,31 @@ pub use ctrl7_g::*;
 pub use ctrl8_xl::*;
 pub use ctrl9_xl::*;
 pub use ctrl10_c::*;
+pub use drd_src::*;
 pub use emb_func_status::*;
 pub use fifo_ctrl::*;
 pub use fifo_data_out::*;
+pub use fifo_data_out_tag::*;
 pub use fifo_status::*;
 pub use free_fall::*;
+pub use fsm_status::*;
 pub use func_cfg_access::*;
 pub use int_ctrl::*;
 pub use int_dur2::*;
-pub use int_ois::*;
-pub use int_src::*;
 pub use internal_freq_fine::*;
 pub use md_cfg::*;
+pub use mlc_status::*;
 pub use ofs_usr::*;
-pub use out_data::*;
+pub use out_temp::*;
 pub use pin_ctrl::*;
+pub use status_master::*;
 pub use status_reg::*;
 pub use tap_cfg::*;
+pub use tap_src::*;
 pub use tap_ths_6d::*;
 pub use timestamp::*;
 pub use wake_up_dur::*;
+pub use wake_up_src::*;
 pub use wake_up_ths::*;
 
 /// Register addresses.
