+++ src/registers/out_temp.rs
@@ -1,0 +1,3 @@
+use bitfield::bitfield;
+
+// TODO: Temperature data output register (r). L and H registers together express a 16-bit word in two’s complement.
